feat: i18n runtime (I18N-P1/P3) + migrate 6 userbins to tr!() (I18N-P2 partial)
- New libneodos/src/i18n.rs: NltTable, i18n_get/load/init, tr!() macro
- NLT format: magic NLT\0, version=1, u32 offsets, O(n) zero-copy lookup
- File::open fix: prepend \Global\FileSystem\ for Ob namespace compat
- i18n_init() reads Language from Registry (fallback en-US)
- Kernel: ensure_language_default() creates Control\Locale\Language
- New userbin/neolocale: validate, stats, diff, check, create
- New scripts/gen_nlt.py: build NLT files from translation tables
- 14 NLT files generated: 7 en-US + 7 es-ES (7 apps each)
- Migrated to tr!(): corehelp, coretype, coredir, corecopy, neoinit, neoshell
- NLT files auto-included in NE2 image via collect_files()
- scripts/gen_system_hiv.py: add Control\Locale\Language = en-US
- Updated IMPROVEMENTS.md -> IMPROVEMENTS_COMPLETED.md

+++ libnet/src/lib.rs
@@ -268,7 +268,7 @@ fn dns_get_server() -> [u8; 4] {
 
     let mut buf = [0u8; 64];
     let res = libneodos::sys_cm_query_value(key, DNS_SERVER_VALUE, &mut buf);
-    let _ = libneodos::sys_close(key as i32);
+    let _ = libneodos::sys_close(key);
 
     match res {
         Ok(total) if total >= 8 => {
